libosmogsm/l1sap.h: extend ph_rach_ind_param with lqual_cb
The link quality, defined by C/I (Carrier-to-Interference) ratio,
can be computed from the training sequence of each burst, where we
can compare the "ideal" training sequence with the actual training
sequence and then express that in cB (centiBels, dB * 10).

By analogy with both RSSI and ToA, it can be used to filter out
false-positive detections and ghost Access Bursts.

Change-Id: Ie2a66ebd040b61d6daf49e04bf8a84d3d64764ee