utils: share static buffer in osmo_str_to{lower,upper}()
This way we get rid of extra 128 bytes in memory per thread created.
It makes sense to share the buffer since it's same size and it doesn't
make much sense to be using both osmo_str_tolower and osmo_strtoupper at
the same time (usually you either want to move everything to uppercase
or everything to lowerase). In required scenarios, one can still use the
_buf versions.

Change-Id: I032803faa0e27c2efdff1ff276acabab95a8319a