osmo-release.sh: update TODO-RELEASE for non-lib projects too
Projects not containing libraries may also want to contain a
TODO-RELEASE in order to write down when a new API available only on
libosmoXYZ current master (hence configure.ac and debian cannot be
updated during the patch using the API until the APIs are available in a
new release). This way, during release process of the project, the
maintainer can see that a release of libosmoXYZ is needed beforehand and
then update configure.ac of project accordingly with the new version of
libosmoXYZ.

Furthermore, we want to update the file only if mode DRY_RUN is not
selected.

Change-Id: I409b7eb8c23d21473f25dd2000f5d4447b24adb9