osmo_fsm: Ensure all state and event names are valid identifiers
we call osmo_identifier_valid() for the FSM name and FSM instance, but we
forgot to do so for all the state and event names.  This meant that they
could contain spaces and the like, which in turn might create problems
when exposing FSM state over CTRL.

This patch shouldn't be merged as-is right now.  We first have to
use it to determine which of our existing programs have related issues.

Change-Id: If98587eff3c48a66ed2e5cc1f01a12accab5a3e7
Closes: OS#4149